refractor: DiplomaMaking -> Move template reading logic to dataHelper
