+++ frontend/src/pages/DiplomaMaking.tsx
@@ -22,6 +22,7 @@ import { SaveButton, SaveButtonType } from "../components/MenuItems/Buttons/Save
 import { AlertPopup, PopupType } from "../components/MenuItems/Popups/AlertPopup";
 import { saltDefaultData } from "../data/data";
 import { getTemplate, makeTemplateInput } from "../templates/baseTemplate";
+import { createTempateInputsFromSaltData, mapInputsToTemplateViewer, mapTemplateInputsToTemplateViewer, tempateInputsFromSaltData, templateInputsFromSaltData } from "../util/dataHelpers";
 
 type Props = {
   bootcamps: BootcampResponse[] | null;
@@ -82,73 +83,8 @@ export default function DiplomaMaking({ bootcamps, templates, addMultipleDiploma
   // When Page Changes -> Loads into PDF preview
   useEffect(() => {
     if(saltData){
-      const inputs = 
-        [makeTemplateInput(
-          populateField(
-            // @ts-ignore
-            saltData[selectedBootcampIndex].template.intro,
-            saltData[selectedBootcampIndex].classname,
-            saltData[selectedBootcampIndex].dategraduate,
-            saltData[selectedBootcampIndex].names[currentPageIndex]
-          ),
-          populateField(
-            // @ts-ignore
-            saltData[selectedBootcampIndex].template.main,
-            saltData[selectedBootcampIndex].classname,
-            saltData[selectedBootcampIndex].dategraduate,
-            saltData[selectedBootcampIndex].names[currentPageIndex]
-          ),
-          populateField(
-            // @ts-ignore
-            saltData[selectedBootcampIndex].template.footer,
-            saltData[selectedBootcampIndex].classname,
-            saltData[selectedBootcampIndex].dategraduate,
-            saltData[selectedBootcampIndex].names[currentPageIndex]
-          ),
-          // @ts-ignore
-          saltData[selectedBootcampIndex].template.basePdf
-        )];
-        const template: Template = getTemplate(
-          inputs[0],
-          { 
-            x: saltData[selectedBootcampIndex].template.introStyling?.xPos ?? null, 
-            y: saltData[selectedBootcampIndex].template.introStyling?.yPos ?? null
-          }, // headerPos
-          { 
-            width: saltData[selectedBootcampIndex].template.introStyling?.width ?? null, 
-            height: saltData[selectedBootcampIndex].template.introStyling?.height ?? null
-          }, // headerSize
-          saltData[selectedBootcampIndex].template.introStyling?.fontSize ?? null, // footerFontSize
-          saltData[selectedBootcampIndex].template.introStyling?.fontColor ?? null, // footerFontColor
-          saltData[selectedBootcampIndex].template.introStyling?.fontName ?? null, // footerFont
-          saltData[selectedBootcampIndex].template.introStyling?.alignment ?? null, // footerAlignment
-          { 
-            x: saltData[selectedBootcampIndex].template.mainStyling?.xPos ?? null, 
-            y: saltData[selectedBootcampIndex].template.mainStyling?.yPos ?? null
-          }, // mainPos
-          { 
-            width: saltData[selectedBootcampIndex].template.mainStyling?.width ?? null, 
-            height: saltData[selectedBootcampIndex].template.mainStyling?.height ?? null
-          }, // mainSize
-          saltData[selectedBootcampIndex].template.mainStyling?.fontSize ?? null, // footerFontSize
-          saltData[selectedBootcampIndex].template.mainStyling?.fontColor ?? null, // footerFontColor
-          saltData[selectedBootcampIndex].template.mainStyling?.fontName ?? null, // footerFont
-          saltData[selectedBootcampIndex].template.mainStyling?.alignment ?? null, // footerAlignment
-          { 
-            x: saltData[selectedBootcampIndex].template.footerStyling?.xPos ?? null, 
-            y: saltData[selectedBootcampIndex].template.footerStyling?.yPos ?? null 
-          }, // footerPos
-          { 
-            width: saltData[selectedBootcampIndex].template.footerStyling?.width ?? null, 
-            height: saltData[selectedBootcampIndex].template.footerStyling?.height ?? null
-          }, // footerSize
-          saltData[selectedBootcampIndex].template.footerStyling?.fontSize ?? null, // footerFontSize
-          saltData[selectedBootcampIndex].template.footerStyling?.fontColor ?? null, // footerFontColor
-          saltData[selectedBootcampIndex].template.footerStyling?.fontName ?? null, // footerFont
-          saltData[selectedBootcampIndex].template.footerStyling?.alignment ?? null // footerAlignment
-        );
-
-        console.log(saltData[selectedBootcampIndex]);
+      const inputs = templateInputsFromSaltData(saltData, selectedBootcampIndex, currentPageIndex);
+      const template = mapTemplateInputsToTemplateViewer(saltData, selectedBootcampIndex, inputs)
 
       getFontsData().then((font) => {
         if (uiRef.current) {
